requirements.txt
aiomysql==0.3.2
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.0
colorama==0.4.6
databases==0.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.1
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
passlib==1.7.4
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
